requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
google-generativeai==0.3.2
reportlab==4.0.4
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
markupsafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
six==1.16.0